Re-subsetting data so there's less missing data
Modified Logan's code so that the data are subsetted to have smaller proportions of missing data. I.e., it seems like the covid and vaccines data were daily and the variants data were biweekly, so I only included covid and vaccines data corresponding with days that we have variants data for. There was also only variants data beginning around January 2021, so the data is from January - November 2021.